Employee
Design a class Employee of an organization. An employee has a name, empid, and salary. Write the default constructor, a constructor with parameters (name, empid, and salary), and methods to return name and salary. Also, write a method increaseSalary that raises the employee’s salary by a certain user-specified percentage. Derive a subclass Manager from the Employee. Add an instance variable named department to the Manager class. Supply a test program that uses these classes and methods.