Account
Design a banking app that has an ‘Account’ class to represent a bank account where a customer can deposit and withdraw money. The app should not permit the customer to withdraw money that exceeds the bank balance. Create a custom exception called “NoSufficientFundException” to handle this scenario and show a more meaningful message to its customers.