The number of times ‘a’ appears in your name
Write a program in Java to create a String object. Initialize this object with your name. Find the length of your name using the appropriate String method. Find whether the character ‘a’ is in your name or not; if yes, find the number of times ‘a’ appears in your name. Print locations of occurrences of ‘a’. Try the same for different String objects